+++ src/collective/compoundcriterion/testing.py
@@ -14,7 +14,7 @@ from plone.app.testing import TEST_USER_NAME
 from plone.testing import z2
 
 import collective.compoundcriterion
-import unittest2 as unittest
+import unittest
 
 
 class CollectiveCompoundcriterionLayer(PloneSandboxLayer):